mcelogmon: base implmentation
Skeleton implemenation for mcelog monitoring plugin. The functionality
implemented simply pings the server first to see if it's alive. If not it sends
a notification.

Added functionality to retrieve Memory Machine Check Errors. Will rename
the file to mcelog in the next patch set. Will also include a persistant
option for notifications and do parsing of the records.

TODO in the future - polling thread.
Tail log file for other error types.

Change-Id: If5c57cceccc486919780b07b6a1d5a589b35bbc1
Signed-off-by: Maryam Tahhan <maryam.tahhan@intel.com>